requirements.txt
torch
transformers
datasets
tokenizers
wandb
pyyaml
trl
